requirements.txt
pyqt5
matplotlib
protobuf
pyserial